Remove unused files and move files around
This commit moves the primary logic for the analytics reporter into the root and renames it `index.js`. The `config.js` and `analytics.js` are moved into a `src` directory and the filenames in the `require` statements are updated.

Additionally, this commit deletes the files in the `app` directory because after reviewing them briefly, I would not expect them to actually work.
