--- app/models.js
@@ -1,6 +1,0 @@
-// Connect to the database.
-var db = {};
-
-module.exports = {
-    data: db
-};
