+++ bin/analytics
@@ -18,7 +18,7 @@
  */
 
 const minimist = require("minimist");
-const run = require("./analytics.js").run;
+const run = require("../index.js").run;
 const options = minimist(process.argv.slice(2));
 
 run(options);
