+++ test/sanity.js
@@ -1,6 +1,6 @@
 
-// make sure analytics.js is free of syntax errors
-require("../analytics");
+// make sure index.js is free of syntax errors
+require("../index");
 
 // one test case beyond that
 require('assert').ok( 1, 'sanity' );
